+++ src/utils/massPathResolve.js
@@ -1,0 +1,3 @@
+import { resolve } from 'node:path';
+
+export default (paths) => paths.map((p) => resolve(p));
